requirements/requirements.txt
# Required Python packages for Deep Learning Models

# Deep Learning Frameworks
tensorflow>=2.13.0
keras>=2.13.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
pytesseract>=0.3.10

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Web Framework
Flask>=2.3.0
werkzeug>=2.3.0

# Utilities
python-dateutil>=2.8.2
